model-server/requirements.txt
# --------- pytorch --------- #
# torch>=1.10.0
# torchvision>=0.11.0
pytorch-lightning==1.7.1
torchmetrics==0.9.3

# --------- hydra --------- #
hydra-core==1.2.0
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- others --------- #
pandas          # for pandas dataframe
transformers    # transformers
pyrootutils     # standardizing the project root setup

fastapi
pydantic
uvicorn
